+++ resources/languages/eng.ini
@@ -1,0 +1,13 @@
+; English language pack for CustomMessages
+
+; Variables supported you can use for your custom messages.
+; {PLAYER} - Display the user's username on the custom message.
+
+; Join Message
+join.message = "&l&6{PLAYER} &fHas joined the server"
+
+; Quit Message
+quit.message = "&l&6{PLAYER} &fHas left the server"
+
+; Whitelisted Message
+whitelistedserver.message = "&l&6This Server is whitelisted right now"
