+++ resources/languages/vie.ini
@@ -1,0 +1,13 @@
+; Vietnamese language pack for CustomMessages
+
+; Variables supported you can use for your custom messages.
+; {PLAYER} - Display the user's username on the custom message.
+
+; Join Message
+join.message = "&l&6{PLAYER} &fĐã tham gia vào máy chủ"
+
+; Quit Message
+quit.message = "&l&6{PLAYER} &fĐã rời khỏi máy chủ"
+
+; Whitelisted Message
+whitelistedserver.message = "&l&6Máy chủ này hiện đã được đưa vào danh sách trắng"
